+++ packages/core/types/index.d.ts
@@ -1,5 +1,6 @@
 import { VueI18n, VueRouter } from './externals';
 
+export * from './boot';
 export * from './externals';
 export * from './main';
 export * from './rx';